Updates CLI so screen is cleared after every prompt
